move ethnicity to fhir converter
