+++ backend/src/main/java/gov/cdc/usds/simplereport/api/MappingConstants.java
@@ -11,9 +11,10 @@ public class MappingConstants {
   public static final String UNKNOWN_STRING = "unknown";
   public static final String U_CODE = "U";
   public static final String ASKED_BUT_UNKNOWN_CODE = "ASKU";
-
   public static final String RACE_EXTENSION_URL =
       "http://ibm.com/fhir/cdm/StructureDefinition/local-race-cd";
-
   public static final String RACE_CODING_SYSTEM = "http://terminology.hl7.org/CodeSystem/v3-Race";
+  public static final String ETHNICITY_EXTENSION_URL =
+      "http://hl7.org/fhir/us/core/StructureDefinition/us-core-ethnicity";
+  public static final String ETHNICITY_CODE_SYSTEM = "urn:oid:2.16.840.1.113883.6.238";
 }
