+++ backend/src/main/java/gov/cdc/usds/simplereport/api/converter/FhirConverter.java
@@ -1,5 +1,8 @@
 package gov.cdc.usds.simplereport.api.converter;
 
+import static gov.cdc.usds.simplereport.api.MappingConstants.ETHNICITY_CODE_SYSTEM;
+import static gov.cdc.usds.simplereport.api.MappingConstants.ETHNICITY_EXTENSION_URL;
+import static gov.cdc.usds.simplereport.api.MappingConstants.NULL_CODE_SYSTEM;
 import static gov.cdc.usds.simplereport.api.MappingConstants.RACE_CODING_SYSTEM;
 import static gov.cdc.usds.simplereport.api.MappingConstants.RACE_EXTENSION_URL;
 
@@ -20,6 +23,7 @@ import java.util.UUID;
 import org.apache.commons.lang3.StringUtils;
 import org.hl7.fhir.r4.model.Address;
 import org.hl7.fhir.r4.model.CodeableConcept;
+import org.hl7.fhir.r4.model.Coding;
 import org.hl7.fhir.r4.model.ContactPoint;
 import org.hl7.fhir.r4.model.ContactPoint.ContactPointSystem;
 import org.hl7.fhir.r4.model.ContactPoint.ContactPointUse;
@@ -28,6 +32,7 @@ import org.hl7.fhir.r4.model.Extension;
 import org.hl7.fhir.r4.model.HumanName;
 import org.hl7.fhir.r4.model.Identifier;
 import org.hl7.fhir.r4.model.Identifier.IdentifierUse;
+import org.hl7.fhir.r4.model.StringType;
 
 public class FhirConverter {
   public static Identifier convertToIdentifier(UUID id) {
@@ -157,18 +162,52 @@ public class FhirConverter {
       if (PersonUtils.raceMap.containsKey(race)) {
         if (MappingConstants.UNKNOWN_STRING.equalsIgnoreCase(race)
             || "refused".equalsIgnoreCase(race)) {
-          coding.setSystem(MappingConstants.NULL_CODE_SYSTEM);
+          coding.setSystem(NULL_CODE_SYSTEM);
         } else {
           coding.setSystem(RACE_CODING_SYSTEM);
         }
         coding.setCode(PersonUtils.raceMap.get(race));
         codeable.setText(race);
       } else {
-        coding.setSystem(MappingConstants.NULL_CODE_SYSTEM);
+        coding.setSystem(NULL_CODE_SYSTEM);
         coding.setCode(MappingConstants.UNK_CODE);
         codeable.setText(MappingConstants.UNKNOWN_STRING);
       }
       ext.setValue(codeable);
+      return ext;
+    }
+    return null;
+  }
+
+  public static Extension convertToEthnicityExtension(String ethnicity) {
+    if (StringUtils.isNotBlank(ethnicity)) {
+      var ext = new Extension();
+      ext.setUrl(ETHNICITY_EXTENSION_URL);
+      var ombExtension = ext.addExtension();
+      ombExtension.setUrl("ombCategory");
+      var ombCoding = new Coding();
+      if (PersonUtils.ETHNICITY_MAP.containsKey(ethnicity)) {
+        if ("refused".equalsIgnoreCase(ethnicity)) {
+          ombCoding.setSystem(NULL_CODE_SYSTEM);
+        } else {
+          ombCoding.setSystem(ETHNICITY_CODE_SYSTEM);
+        }
+        ombCoding.setCode(PersonUtils.ETHNICITY_MAP.get(ethnicity).get(0));
+        ombCoding.setDisplay(PersonUtils.ETHNICITY_MAP.get(ethnicity).get(1));
+
+        var text = ext.addExtension();
+        text.setUrl("text");
+        text.setValue(new StringType(PersonUtils.ETHNICITY_MAP.get(ethnicity).get(1)));
+      } else {
+        ombCoding.setSystem(NULL_CODE_SYSTEM);
+        ombCoding.setCode(MappingConstants.UNK_CODE);
+        ombCoding.setDisplay(MappingConstants.UNKNOWN_STRING);
+
+        var text = ext.addExtension();
+        text.setUrl("text");
+        text.setValue(new StringType(MappingConstants.UNKNOWN_STRING));
+      }
+      ombExtension.setValue(ombCoding);
       return ext;
     }
     return null;
